// legacy-structure/client/src/modules/finance/models/models.ts
export interface OptionFilter {
  id: number;
  title: string;
  label: string;
  isCollaborator?: boolean;
}

export interface FilterIssuesModel {
  milestones: null | OptionFilter;
  labels: null | OptionFilter;
  project: null | OptionFilter;
  assignedTo: null | OptionFilter;
  repository: null | OptionFilter;
  status: null | OptionFilter;
  minDate: null | string;
  maxDate: null | string;
}
// ...
export interface IssuesRecord {
  assignedTo: {
    login: string;
  };
  description: string;
  assignedToId: number;
  number: number;
  milestones: string;
  labels: string;
  last_comment: string;
  created_at: string;
  closed_at: string;
  project: {
    title: string;
    id: number;
  };
  repository: {
    title: string;
    id: number;
  };
  id: number | '';
  title: string;
  status: string;
  hours: number | '';
  url: string;
  accountId: number;
}
// ...
export interface OrganizationDataRecord {
  name: string;
  description: string;
  address: string;
  place: string;
  phone: string;
  cap: number;
  email: string;
  web: string;
  no: string;
  accountingInternetPrice: number;
  accountingUnit: string;
  accountingCoin: string;
  accountingLabel: string;
  bankName: string;
  bankBIC: string;
  bankIBAN: string;
  bankNo: string;
  bankCap: string;
  bankPlace: string;
  bankAddress: string;
  bankValidDays: number;
}